Squashed 'pysc/' changes from 276a1d8..63f58cc
63f58cc Merging local changes with TUDO tree
a4e0290 [arm] Initial ARM Hello-World ramp-up.
e098ccf Removing prints
75a9551 Use arem in quadcopter
62d3c3b Getting the quadcopter platform to work
53e1608 Correcting clkdevice and memdevice path
2864c32 splitting core/common in base, common, utils and amba
8c514c3 Windows preperation
2c2bc32 Fixing SoCRocket for subtree and waf 1.91
f870eb6 Renaming usi->get_ifs_tuple to usi->__usi_interfaces__

git-subtree-dir: pysc
git-subtree-split: 63f58cc5db8383eba95a7e50bf47ab472f3f086e

+++ usi/api/systemc.h
@@ -13,7 +13,7 @@
 #define PYSC_API_SYSTEMC_H
 
 #include <Python.h>
-#include "core/common/sc_api.h"
+#include "core/base/sc_api.h"
 
 #ifdef SWIG
 namespace sc_core {
@@ -38,7 +38,7 @@ double delta_count();
 
 int set_verbosity(int verbosity);
 
-long get_curr_process_handle();
+size_t get_curr_process_handle();
 bool is_running();
 std::vector<sc_core::sc_object *> get_top_level_objects();
